+++ src/snowflake/snowpark/dataframe_writer.py
@@ -1,9 +1,9 @@
 #
-# Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
+# Copyright (c) 2012-2024 Snowflake Computing Inc. All rights reserved.
 #
 
 import sys
-from typing import Dict, List, Literal, Optional, Union, overload
+from typing import Any, Dict, List, Literal, Optional, Union, overload
 
 import snowflake.snowpark  # for forward references of type hints
 from snowflake.snowpark._internal.analyzer.snowflake_plan_node import (
@@ -11,6 +11,7 @@ from snowflake.snowpark._internal.analyzer.snowflake_plan_node import (
     SaveMode,
     SnowflakeCreateTable,
 )
+from snowflake.snowpark._internal.open_telemetry import open_telemetry_context_manager
 from snowflake.snowpark._internal.telemetry import (
     add_api_call,
     dfw_collect_api_telemetry,
@@ -167,60 +168,65 @@ class DataFrameWriter:
             >>> session.table("my_transient_table").collect()
             [Row(A=1, B=2), Row(A=3, B=4)]
         """
-        save_mode = (
-            str_to_enum(mode.lower(), SaveMode, "'mode'") if mode else self._save_mode
-        )
-        full_table_name = (
-            table_name if isinstance(table_name, str) else ".".join(table_name)
-        )
-        validate_object_name(full_table_name)
-        table_name = (
-            parse_table_name(table_name) if isinstance(table_name, str) else table_name
-        )
-        if column_order is None or column_order.lower() not in ("name", "index"):
-            raise ValueError("'column_order' must be either 'name' or 'index'")
-        column_names = (
-            self._dataframe.columns if column_order.lower() == "name" else None
-        )
-        clustering_exprs = (
-            [
-                _to_col_if_str(col, "DataFrameWriter.save_as_table")._expression
-                for col in clustering_keys
-            ]
-            if clustering_keys
-            else []
-        )
-
-        if create_temp_table:
-            warning(
-                "save_as_table.create_temp_table",
-                "create_temp_table is deprecated. We still respect this parameter when it is True but "
-                'please consider using `table_type="temporary"` instead.',
+        with open_telemetry_context_manager(self.save_as_table, self._dataframe):
+            save_mode = (
+                str_to_enum(mode.lower(), SaveMode, "'mode'")
+                if mode
+                else self._save_mode
             )
-            table_type = "temporary"
-
-        if table_type and table_type.lower() not in SUPPORTED_TABLE_TYPES:
-            raise ValueError(
-                f"Unsupported table type. Expected table types: {SUPPORTED_TABLE_TYPES}"
+            full_table_name = (
+                table_name if isinstance(table_name, str) else ".".join(table_name)
+            )
+            validate_object_name(full_table_name)
+            table_name = (
+                parse_table_name(table_name)
+                if isinstance(table_name, str)
+                else table_name
+            )
+            if column_order is None or column_order.lower() not in ("name", "index"):
+                raise ValueError("'column_order' must be either 'name' or 'index'")
+            column_names = (
+                self._dataframe.columns if column_order.lower() == "name" else None
+            )
+            clustering_exprs = (
+                [
+                    _to_col_if_str(col, "DataFrameWriter.save_as_table")._expression
+                    for col in clustering_keys
+                ]
+                if clustering_keys
+                else []
             )
 
-        create_table_logic_plan = SnowflakeCreateTable(
-            table_name,
-            column_names,
-            save_mode,
-            self._dataframe._plan,
-            table_type,
-            clustering_exprs,
-        )
-        session = self._dataframe._session
-        snowflake_plan = session._analyzer.resolve(create_table_logic_plan)
-        result = session._conn.execute(
-            snowflake_plan,
-            _statement_params=statement_params or self._dataframe._statement_params,
-            block=block,
-            data_type=_AsyncResultType.NO_RESULT,
-        )
-        return result if not block else None
+            if create_temp_table:
+                warning(
+                    "save_as_table.create_temp_table",
+                    "create_temp_table is deprecated. We still respect this parameter when it is True but "
+                    'please consider using `table_type="temporary"` instead.',
+                )
+                table_type = "temporary"
+
+            if table_type and table_type.lower() not in SUPPORTED_TABLE_TYPES:
+                raise ValueError(
+                    f"Unsupported table type. Expected table types: {SUPPORTED_TABLE_TYPES}"
+                )
+
+            create_table_logic_plan = SnowflakeCreateTable(
+                table_name,
+                column_names,
+                save_mode,
+                self._dataframe._plan,
+                table_type,
+                clustering_exprs,
+            )
+            session = self._dataframe._session
+            snowflake_plan = session._analyzer.resolve(create_table_logic_plan)
+            result = session._conn.execute(
+                snowflake_plan,
+                _statement_params=statement_params or self._dataframe._statement_params,
+                block=block,
+                data_type=_AsyncResultType.NO_RESULT,
+            )
+            return result if not block else None
 
     @overload
     def copy_into_location(
@@ -233,8 +239,8 @@ class DataFrameWriter:
         format_type_options: Optional[Dict[str, str]] = None,
         header: bool = False,
         statement_params: Optional[Dict[str, str]] = None,
-        block: bool = True,
-        **copy_options: Optional[str],
+        block: Literal[True] = True,
+        **copy_options: Optional[Dict[str, Any]],
     ) -> List[Row]:
         ...  # pragma: no cover
 
@@ -249,8 +255,8 @@ class DataFrameWriter:
         format_type_options: Optional[Dict[str, str]] = None,
         header: bool = False,
         statement_params: Optional[Dict[str, str]] = None,
-        block: bool = False,
-        **copy_options: Optional[str],
+        block: Literal[False] = False,
+        **copy_options: Optional[Dict[str, Any]],
     ) -> AsyncJob:
         ...  # pragma: no cover
 
@@ -265,7 +271,7 @@ class DataFrameWriter:
         header: bool = False,
         statement_params: Optional[Dict[str, str]] = None,
         block: bool = True,
-        **copy_options: Optional[str],
+        **copy_options: Optional[Dict[str, Any]],
     ) -> Union[List[Row], AsyncJob]:
         """Executes a `COPY INTO <location> <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html>`__ to unload data from a ``DataFrame`` into one or more files in a stage or external stage.
 
@@ -335,34 +341,144 @@ class DataFrameWriter:
             block=block,
         )
 
-    def csv(self, path: str, overwrite: bool = False, compression: str = None, single: bool = True,
-            partition_by: ColumnOrName = None) -> Union[List[Row], AsyncJob]:
+    def csv(
+        self,
+        location: str,
+        *,
+        partition_by: Optional[ColumnOrSqlExpr] = None,
+        format_type_options: Optional[Dict[str, str]] = None,
+        header: bool = False,
+        statement_params: Optional[Dict[str, str]] = None,
+        block: bool = True,
+        **copy_options: Optional[str],
+    ) -> Union[List[Row], AsyncJob]:
         """Executes internally a `COPY INTO <location> <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html>`__ to unload data from a ``DataFrame`` into one or more CSV files in a stage or external stage.
 
-                Args:
-                    path: The destination stage location.
-                    overwrite: Specifies if it should overwrite the file if exists, the default value is ``False``.
-                    compression: String (constant) that specifies to compresses the unloaded data files using the specified compression algorithm. Use the options documented in the `Format Type Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#format-type-options-formattypeoptions>`__
-                    single: Boolean that specifies whether to generate a single file or multiple files. If FALSE, a filename prefix must be included in ``<path>``
-                    partition_by: Specifies an expression used to partition the unloaded table rows into separate files. It can be a :class:`Column`, a column name, or a SQL expression.
+        Args:
+            location: The destination stage location.
+            partition_by: Specifies an expression used to partition the unloaded table rows into separate files. It can be a :class:`Column`, a column name, or a SQL expression.
+            format_type_options: Depending on the ``file_format_type`` specified, you can include more format specific options. Use the options documented in the `Format Type Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#format-type-options-formattypeoptions>`__.
+            header: Specifies whether to include the table column headings in the output files.
+            statement_params: Dictionary of statement level parameters to be set while executing this action.
+            copy_options: The kwargs that are used to specify the copy options. Use the options documented in the `Copy Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#copy-options-copyoptions>`__.
+            block: A bool value indicating whether this function will wait until the result is available.
+                When it is ``False``, this function executes the underlying queries of the dataframe
+                asynchronously and returns an :class:`AsyncJob`.
+        Returns:
+            A list of :class:`Row` objects containing unloading results.
 
-                Returns:
-                    A list of :class:`Row` objects containing unloading results.
+        Example::
 
-                Example::
+            >>> # save this dataframe to a csv file on the session stage
+            >>> df = session.create_dataframe([["John", "Berry"], ["Rick", "Berry"], ["Anthony", "Davis"]], schema = ["FIRST_NAME", "LAST_NAME"])
+            >>> remote_file_path = f"{session.get_session_stage()}/names.csv"
+            >>> copy_result = df.write.csv(remote_file_path, overwrite=True, single=True)
+            >>> copy_result[0].rows_unloaded
+            3
+        """
+        return self.copy_into_location(
+            location,
+            file_format_type="CSV",
+            partition_by=partition_by,
+            format_type_options=format_type_options,
+            header=header,
+            statement_params=statement_params,
+            block=block,
+            **copy_options,
+        )
 
-                    >>> # save this dataframe to a parquet file on the session stage
-                    >>> df = session.create_dataframe([["John", "Berry"], ["Rick", "Berry"], ["Anthony", "Davis"]], schema = ["FIRST_NAME", "LAST_NAME"])
-                    >>> remote_file_path = f"{session.get_session_stage()}/names.csv"
-                    >>> copy_result = df.write.csv(remote_file_path, overwrite=True)
-                    >>> copy_result[0].rows_unloaded
-                    3
-                """
-        return self.copy_into_location(path,
-                                       file_format_type="CSV",
-                                       partition_by=partition_by,
-                                       overwrite=overwrite,
-                                       format_type_options=dict(compression=compression),
-                                       single=single)
+    def json(
+        self,
+        location: str,
+        *,
+        partition_by: Optional[ColumnOrSqlExpr] = None,
+        format_type_options: Optional[Dict[str, str]] = None,
+        header: bool = False,
+        statement_params: Optional[Dict[str, str]] = None,
+        block: bool = True,
+        **copy_options: Optional[str],
+    ) -> Union[List[Row], AsyncJob]:
+        """Executes internally a `COPY INTO <location> <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html>`__ to unload data from a ``DataFrame`` into a JSON file in a stage or external stage.
+
+        Args:
+            location: The destination stage location.
+            partition_by: Specifies an expression used to partition the unloaded table rows into separate files. It can be a :class:`Column`, a column name, or a SQL expression.
+            format_type_options: Depending on the ``file_format_type`` specified, you can include more format specific options. Use the options documented in the `Format Type Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#format-type-options-formattypeoptions>`__.
+            header: Specifies whether to include the table column headings in the output files.
+            statement_params: Dictionary of statement level parameters to be set while executing this action.
+            copy_options: The kwargs that are used to specify the copy options. Use the options documented in the `Copy Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#copy-options-copyoptions>`__.
+            block: A bool value indicating whether this function will wait until the result is available.
+                When it is ``False``, this function executes the underlying queries of the dataframe
+                asynchronously and returns an :class:`AsyncJob`.
+
+        Returns:
+            A list of :class:`Row` objects containing unloading results.
+
+        Example::
+
+            >>> # save this dataframe to a json file on the session stage
+            >>> df = session.sql("select parse_json('[{a: 1, b: 2}, {a: 3, b: 0}]')")
+            >>> remote_file_path = f"{session.get_session_stage()}/names.json"
+            >>> copy_result = df.write.json(remote_file_path, overwrite=True, single=True)
+            >>> copy_result[0].rows_unloaded
+            1
+        """
+        return self.copy_into_location(
+            location,
+            file_format_type="JSON",
+            partition_by=partition_by,
+            format_type_options=format_type_options,
+            header=header,
+            statement_params=statement_params,
+            block=block,
+            **copy_options,
+        )
+
+    def parquet(
+        self,
+        location: str,
+        *,
+        partition_by: Optional[ColumnOrSqlExpr] = None,
+        format_type_options: Optional[Dict[str, str]] = None,
+        header: bool = False,
+        statement_params: Optional[Dict[str, str]] = None,
+        block: bool = True,
+        **copy_options: Optional[str],
+    ) -> Union[List[Row], AsyncJob]:
+        """Executes internally a `COPY INTO <location> <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html>`__ to unload data from a ``DataFrame`` into a PARQUET file in a stage or external stage.
+
+        Args:
+            location: The destination stage location.
+            partition_by: Specifies an expression used to partition the unloaded table rows into separate files. It can be a :class:`Column`, a column name, or a SQL expression.
+            format_type_options: Depending on the ``file_format_type`` specified, you can include more format specific options. Use the options documented in the `Format Type Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#format-type-options-formattypeoptions>`__.
+            header: Specifies whether to include the table column headings in the output files.
+            statement_params: Dictionary of statement level parameters to be set while executing this action.
+            copy_options: The kwargs that are used to specify the copy options. Use the options documented in the `Copy Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#copy-options-copyoptions>`__.
+            block: A bool value indicating whether this function will wait until the result is available.
+                When it is ``False``, this function executes the underlying queries of the dataframe
+                asynchronously and returns an :class:`AsyncJob`.
+
+        Returns:
+            A list of :class:`Row` objects containing unloading results.
+
+        Example::
+
+            >>> # save this dataframe to a parquet file on the session stage
+            >>> df = session.create_dataframe([["John", "Berry"], ["Rick", "Berry"], ["Anthony", "Davis"]], schema = ["FIRST_NAME", "LAST_NAME"])
+            >>> remote_file_path = f"{session.get_session_stage()}/names.parquet"
+            >>> copy_result = df.write.parquet(remote_file_path, overwrite=True, single=True)
+            >>> copy_result[0].rows_unloaded
+            3
+        """
+        return self.copy_into_location(
+            location,
+            file_format_type="PARQUET",
+            partition_by=partition_by,
+            format_type_options=format_type_options,
+            header=header,
+            statement_params=statement_params,
+            block=block,
+            **copy_options,
+        )
 
     saveAsTable = save_as_table